feat: connect buttons to scripts
