+++ packages/nc-gui/components/smartsheet/column/ButtonOptions/Script.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts"></script>
+
+<template></template>
+
+<style scoped lang="scss"></style>
