+++ packages/nc-gui/lang/en.json
@@ -678,6 +678,7 @@
     "limitRecordSelectionToFilters": "Limit record selection to filters",
     "addedBy": "Added by",
     "whatsNew": "What's New!",
+    "runScript": "Run Script",
     "scriptName": "Script name",
     "enableSnowFall": "Enable Snowfall",
     "disableSnowFall": "Disable Snowfall",
@@ -706,7 +707,8 @@
     "editDescription": "Edit description",
     "urlFormula": "URL Formula",
     "selectIcon": "none",
-    "selectAWebhook": "--select a webhook--",
+    "selectAWebhook": "-- select a webhook --",
+    "selectAScript": "-- select a script --",
     "openUrl": "Open URL",
     "runWebHook": "Run Webhook",
     "onClick": "On click",
