+++ packages/nocodb/src/models/Column.ts
@@ -364,6 +364,7 @@ export default class Column<T = any> implements ColumnType {
           theme: column.theme,
           color: column.color,
           fk_webhook_id: column?.fk_webhook_id,
+          fk_script_id: column?.fk_script_id,
           label: column.label,
           fk_integration_id: column.fk_integration_id,
           model: column.model,
